+++ simulator/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to the contest simulator will be documented in this file.
 Dates are based on Eastern Time Zone.
 
+## 2024-12-19
+
+### Added
+
+- Basic support for skill card customization
+
 ## 2024-12-18
 
 ### Changed